scada-simulator/requirements.txt
flask
google-cloud-pubsub
gunicorn